+++ dolly0920/src/main/java/com/example/dolly0920/item7/HashMapTest.java
@@ -1,4 +1,4 @@
-package com.example.dolly0920.item8;
+package com.example.dolly0920.item7;
 
 import java.util.HashMap;
